+++ src/Flows/Student/Student.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import Signin from './Signin/Signin';
 import Dashboard from './Dashboard/Dashboard';
-import Register from './Register/Register';
+import Register from './Register/StudentRegister.jsx';
 import { checkUserExists } from '../../service/checkUserExists.js';
 import { checkUserTypes } from '../../service/checkUserTypes.js';
 
@@ -22,6 +22,7 @@ const Student = () => {
 
 
   // if users loged in 
+  // auto checking for first time
   useEffect(() => {
     const checkUserStatus = async () => {
       const studentUid = localStorage.getItem('student_uid');
@@ -33,15 +34,17 @@ const Student = () => {
         }
       }
     };
-
     checkUserStatus();
   }, []);
 
-  const handleSignin = async() => {
+  const handleSignin = async () => {
+    console.log("handlesignin");
     const studentUid = localStorage.getItem('student_uid');
     const userTypes = await checkUserTypes(studentUid);
     if (userTypes === "team_leader") {
       setIsSignedIn(true);
+      alert("You are a student ");
+      console.log("You are a student ");
     }
     else {
       alert("You are not a student ");
@@ -56,18 +59,17 @@ const Student = () => {
     // you may redirect to home also 
   }
 
-  const toggleRegistering = () => {
-    setIsRegistering(!isRegistering);
-  };
 
   return (
     <div>
       {!isSignedIn && (
         <>
           {isRegistering ? (
-            <Register onToggle={toggleRegistering} />
+            <Register />
           ) : (
-            <Signin onSignin={handleSignin} onToggle={toggleRegistering} />
+            <Signin
+              onSignin={handleSignin}
+            />
           )}
         </>
       )}
