+++ src/components/FlowComponents/OtpInputContainer/OtpInputContainer.module.css
@@ -1,0 +1,24 @@
+.inputbox {
+  display: flex;
+  justify-content: space-between;
+}
+
+.inputbox input {
+  padding: 1rem 1.2rem;
+  background-color: #000000;
+  border: 1px solid #ffffff4d;
+  border-radius: 5px;
+  color: #8190bf;
+  outline: none;
+  font-size: 20px;
+  width: 25px;
+  height: 25px;
+  max-width: 25px;
+  max-height: 25px;
+  text-align: center;
+  margin-right: 10px;
+}
+
+.inputbox input::placeholder {
+  color: #8190bf;
+}
